requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM API clients
openai>=1.0.0
anthropic>=0.7.0  # Optional: if using Claude

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel output

# Utilities
python-dotenv>=1.0.0  # For loading environment variables
tqdm>=4.65.0  # Progress bars
tenacity>=8.2.0  # Retry logic

# Text processing
html2text>=2020.1.16
readability-lxml>=0.8.1
trafilatura>=1.6.0  # For better content extraction

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional: for async processing (if you want to add async later)
# aiohttp>=3.8.0
# asyncio-throttle>=1.0.0